Do not set balance based on last valid Statement
There seems to be some kind of strange mismatch between documentation of
the plugin interface and HBCI code

The plugin documentation for statements:

    "value" = "123.456,00 EUR", (Umsatz als String, korrekte Konvertierung erfolgt in Pecunia)

However the code uses the value of the last statement as balance of the
account.

Two questions arise:

1. Why should I send a "balance" in the enclosing struct at all? Here the
   plugin documentation:
     "balance": "100,0 EUR", (Saldo als String, korrekte Konvertierung erfolgt in Pecunia)
2. If value should be a \sum why is it not documented?

As a result I hereby use the balance element and call this sufficient for
plugins.